codes/requirements.txt
streamlit==1.29.0
pymongo==4.6.0
python-dotenv==1.0.0
dnspython==2.4.2